fix the data type

+++ src/main/entity/PeriodicInflow.java
@@ -3,7 +3,7 @@ package entity;
 import java.time.LocalDate;
 
 public class PeriodicInflow extends PeriodicTransaction {
-    public PeriodicInflow(String identification, double amount, LocalDate date, String description, int recurrencePeriodInDays) {
+    public PeriodicInflow(String identification, float amount, LocalDate date, String description, int recurrencePeriodInDays) {
         super(identification, amount, date, description, recurrencePeriodInDays);
         if(!this.isInflow()) System.out.println("Periodic inflow error, should be outflow!");
     }
