+++ src/main/entity/PeriodicOutflow.java
@@ -3,7 +3,7 @@ package entity;
 import java.time.LocalDate;
 
 public class PeriodicOutflow extends PeriodicTransaction {
-    public PeriodicOutflow(String identification, double amount, LocalDate date, String description, int recurrencePeriodInDays) {
+    public PeriodicOutflow(String identification, float amount, LocalDate date, String description, int recurrencePeriodInDays) {
         super(identification, amount, date, description, recurrencePeriodInDays);
         if(this.isInflow()) System.out.println("Periodic outflow error, should be inflow!");
     }
